added authorization too to the routes, so only the user who created a resourche can change or delete is

+++ controllers/feed.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-underscore-dangle */
 /* eslint-disable no-param-reassign */
 const fs = require('fs');
 const path = require('path');
@@ -5,6 +6,7 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -19,13 +21,11 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage);
     })
     .then(posts => {
-      res
-        .status(200)
-        .json({
-          message: 'Fetched posts successfully.',
-          posts: posts,
-          totalItems: totalItems
-        });
+      res.status(200).json({
+        message: 'Fetched posts successfully.',
+        posts: posts,
+        totalItems: totalItems
+      });
     })
     .catch(err => {
       const newError = err;
@@ -52,20 +52,29 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const { title } = req.body;
   const { content } = req.body;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Lacen'
-    }
+    creator: req.userId
   });
   post
     .save()
+    // eslint-disable-next-line arrow-body-style
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
     .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -125,6 +134,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -137,11 +151,10 @@ exports.updatePost = (req, res, next) => {
       res.status(200).json({ message: 'Post updated!', post: result });
     })
     .catch(err => {
-      const newError = err;
-      if (!newError.status) {
-        newError.statusCode = 500;
+      if (!err.statusCode) {
+        err.statusCode = 500;
       }
-      next(newError);
+      next(err);
     });
 };
 
@@ -154,7 +167,11 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      // check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
@@ -162,11 +179,10 @@ exports.deletePost = (req, res, next) => {
       res.status(200).json({ message: 'Deleted post.' });
     })
     .catch(err => {
-      const newError = err;
-      if (!newError.status) {
-        newError.statusCode = 500;
+      if (!err.statusCode) {
+        err.statusCode = 500;
       }
-      next(newError);
+      next(err);
     });
 };
